ai-exam-backend/requirements.txt
fastapi
uvicorn
supabase
google-generativeai
requests
pydantic
python-dotenv
python-multipart